Fixed a bug while saving the field
If the user omitted to select a value for the section in one field, all other fields lost their settings for the allowed sections. This made saving after a error create more error.

The problem is that we discarded values when it was already an array. Since PHP is so dynamic, we need to handle all cases: string, null and array.

cc @PascalPiche