Better error management + memento
The memento implement reverts the field value to its original, pre-save
value.

I choose this behavior because that's what the data still is in the
database. So we always show the current, saved state.

Also, the save method is called each time an atomic operation is made so
the risk of loosing lots of work is minimal.