Logic clean up
The old data structure was prone to error since it needed a lot of hacks and check for multiple situations.

This commits put the axe in this by creating more robust data structures that should avoid nasty if conditions.

It fixes bug a lot, so developers that built upon those bugs will need to change their field/data-source settings or update their xsl markup.

Also this should fix #55