Fixed errors and memento overwrite then min > 0
This commits adds a check before doing the ajax save to make sure we have enough items.

The field (obviously) does not validate when the minimum number is not reached, but we must let the user select the entries beforehand. Without this, editing existing entries with 0 item and a minimum of 4 was impossible.

cc @beaubbe @pascalpiche